refactor(eval): consolidate harness into one fixture file
Merge the agent fixture, matchers, judge, config isolation, and result
recording into eval/it.ts; delete claude.ts, report.ts, and setup.ts.

Move the judge ratchet from a read-back of results.jsonl into an inline
threshold literal that vitest -u rewrites like a snapshot, so results.jsonl
is write-only. Trim recorded fields to effectiveness and effort (turns,
tokens, duration); drop cost, which swings with provider pricing.
